requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
jinja2==3.1.3
scikit-learn==1.3.2
python-multipart==0.0.9
requests==2.32.3
httpx==0.27.0
python-dotenv==1.0.1
trafilatura==1.9.0
